--- a/inc/item_rack.class.php
+++ b/inc/item_rack.class.php
@@ -1075,7 +1075,7 @@ JAVASCRIPT;
       return $input;
    }
 
-   function getRawName() {
+   public function computeFriendlyName() {
       $rack = new Rack();
       $rack->getFromDB($this->fields['racks_id']);
       $name = sprintf(