--- a/inc/itiltemplatefield.class.php
+++ b/inc/itiltemplatefield.class.php
@@ -75,7 +75,7 @@ abstract class ITILTemplateField extends CommonDBChild {
    }
 
 
-   function getRawName() {
+   function computeFriendlyName() {
       $tt_class = static::$itemtype;
       $tt     = new $tt_class;
       $fields = $tt->getAllowedFieldsNames(true);