--- a/inc/tickettemplatemandatoryfield.class.php
+++ b/inc/tickettemplatemandatoryfield.class.php
@@ -42,5 +42,4 @@ class TicketTemplateMandatoryField extends ITILTemplateMandatoryField {
    static public $itemtype  = 'TicketTemplate';
    static public $items_id  = 'tickettemplates_id';
    static public $itiltype = 'Ticket';
-
 }