… CHANGELOG.md
… apirest.md
… tests/functionnal/RuleCriteria.php
… tests/functionnal/User.php