Some API requests returns ids (e.g. users_id) that a self-service user can't work with (as he is not allowed to query /User/{id}) and thus can't get the name matching this id (e.g. the name of the author of an ITILFollowup).

This proposal add a "forcename" option that allow to access the rawname of the items.  

Example request: 
```http
GET apirest.php/ITILFollowup?id=22
&forcename[0]=users_id
&forcename[1]=users_id_editor
```

Response:
```json
{
        "id": 22,
        "itemtype": "Ticket",
        "items_id": 5,
        "date": "2019-02-25 16:27:05",
        "users_id": 2,
        "users_id_editor": 2,
        "content": "&lt;p&gt;aaaqdzdqdqzdqzdqzdqz&lt;/p&gt;&lt;p&gt;&lt;a href=\"http://72.glpi-fusion-94.local/front/document.send.php?docid=4&amp;tickets_id=5\" target=\"_blank\" rel=\"noopener\"&gt;&lt;img src=\"http://72.glpi-fusion-94.local/front/document.send.php?docid=4&amp;tickets_id=5\" alt=\"3577bb62-c6a01f19-5c740943affc40.83487561\" width=\"185\" /&gt;&lt;/a&gt;&lt;/p&gt;",
        "is_private": 0,
        "requesttypes_id": 1,
        "date_mod": "2020-01-28 10:06:09",
        "date_creation": "2019-02-25 16:27:05",
        "timeline_position": 1,
        "sourceitems_id": 0,
        "sourceof_items_id": 0,
 |      "_names": {
 |          "users_id": "glpi",
 |          "users_id_editor": "glpi",
 |          "links": [
 |              {
 |                  "rel": "User",
 |                  "href": "http://glpi-fusion-94.local/api/User/glpi"
 |              },
 |              {
 |                  "rel": "User",
 |                  "href": "http://glpi-fusion-94.local/api/User/glpi"
 |              }
 |          ]
 |      }
}
```


Performance impact:

Benchmark request (80 followups, two names requested):
```http
GET apirest.php/ITILFollowup?range=0-80
&forcename[0]=users_id
&forcename[1]=users_id_editor
```

Without cache:
```
PHP message: Additional time: 76.05 ms
PHP message: Average time per name fetched:  0.93 ms
```

With cache:
```
PHP message: Additional time: 30.44 ms
PHP message: Average time per name fetched:  0.37 ms
```

| Q             | A
| ------------- | ---
| Bug fix?      | no
| New feature?  | yes
| BC breaks?    | no
| Deprecations? | no
| Tests pass?   | yes