@AdrienClairembault 2020-01-31
@cedric-anne 

As we discussed yesterday I agree with you on this except on one point:
> makes CommonDBTM::getFriendlyName() use cache when it is available or call CommonDBTM::computeFriendlyName() to init its cache when needed.

CommonDBTM::getFriendlyName() should set the cache but not use it.
Building the friendly name has no performance cost, what slow down the API is to load the object from DB before being able to get his friendly name.

This mean that the API should still check the cache to avoid loading the object if his friendly name is cached (and CommonDBTM::getFriendlyName should set the friendly name cache so the API can benefit from already cached names even if it was not the one who used them).