@AdrienClairembault 2020-02-04
> Seems OK for me, but should it target 9.4 or 9.5 branch ?

Rebased to 9.5.