@cedric-anne 2020-02-04
An entry in CHANGELOG.md should be added for deprecation of `getRawName()`.