@AdrienClairembault 2020-02-10
As we discussed @orthagh @cedric-anne :
* forcename => add_keys_names
* _names => _keys_names
* self =-> id