@rafaeldomi 2020-02-25
Trying to install a freshly cloned 9.5/bugfixes i get this error:

GLPI SETUP

Step 3
Initialization of the database
Uncaught Exception Error: Call to a member function delete() on null in [..]/glpi95/inc/commondbtm.class.php at line 1455, which points to this line, commited in this pull

```
            // Update raw names cache (for API)
            $GLPI_CACHE->delete(self::getCacheKeyForFriendlyName(
               $this->getType(),
               $this->fields['id']
            ));
            $this->getFriendlyName();
```